SQUASH???
Let's declare that for the purpose of "pull --quiet", the new "you
might have created a merge when you may wanted to rebase" warning is
an unwanted noise.

Revert the hack made to t5521; the second test that creates a new
repository and runs a pull inside it will not get affected by the
local configuration setting it makes to the original trash
repository and the breakage won't be worked around with it anyways.

This is not a full remedy for the previous step, yet.  The topic
needs a test to ensure the warning is emitted when it should be.
Existing t5521 is a test that ensures the warning is not given in
one of the conditions it should not be (i.e. when --quiet is in
effect).  Negative tests for other cases (e.g. when --ff-only is
given from the command line, without pull.rebase configured) should
also be added.